[fix] MemberUseCase 클래스 중복 수정
